rename to TlmCmd from CmdTlm
