+++ my_mod/cmd_def.py
@@ -8,7 +8,7 @@ import sys
 import pprint
 
 def GenerateCmdDef(settings, sgc_db):
-    output_file_path = settings["c2a_root_dir"] + r"src_user/CmdTlm/"
+    output_file_path = settings["c2a_root_dir"] + r"src_user/TlmCmd/"
     output_file_name_base = "command_definitions"
 
     DATA_SART_ROW = 3
@@ -81,7 +81,7 @@ def GenerateCmdDef(settings, sgc_db):
 
 
 def GenerateBctDef(settings, bct_db):
-    output_file_path = settings["c2a_root_dir"] + r"src_user/CmdTlm/"
+    output_file_path = settings["c2a_root_dir"] + r"src_user/TlmCmd/"
     output_file_name = "block_command_definitions.h"
 
     DATA_SART_ROW = 2
@@ -160,7 +160,7 @@ def OutputCmdDefC_(file_path, body):
  * @brief  コマンド定義
  * @note   このコードは自動生成されています！
  */
-#include "../../src_core/CmdTlm/command_analyze.h"
+#include "../../src_core/TlmCmd/command_analyze.h"
 #include "command_definitions.h"
 #include "command_source.h"
