+++ settings.json
@@ -9,7 +9,7 @@
       "name" : "AOBC",
       "is_enable" : 1,
       "db_prefix" : "SAMPLE_AOBC",
-      "db_path" : "C:/Docs/ISSL_Git/S2E/FlightSW/c2a_sample_aobc/src/src_user/Settings/CmdTlm/DataBase/",
+      "db_path" : "C:/Docs/ISSL_Git/S2E/FlightSW/c2a_sample_aobc/src/src_user/Settings/TlmCmd/DataBase/",
       "tlm_max_contents_len" : 512,
       "driver_path" : "Aocs/",
       "driver_type" : "AOBC_Driver",
@@ -21,7 +21,7 @@
       "name" : "TOBC",
       "is_enable" : 1,
       "db_prefix" : "SAMPLE_TOBC",
-      "db_path" : "C:/Docs/ISSL_Git/S2E/FlightSW/c2a_sample_tobc/src/src_user/Settings/CmdTlm/DataBase/",
+      "db_path" : "C:/Docs/ISSL_Git/S2E/FlightSW/c2a_sample_tobc/src/src_user/Settings/TlmCmd/DataBase/",
       "tlm_max_contents_len" : 512,
       "driver_path" : "Thermal/",
       "driver_type" : "TOBC_Driver",
